JavaScript: unify maybeAddImport and bindModule into maybeBind
Collapses the two module-binding APIs into one: maybeBind serves both the
AMD and ESM/CommonJS lanes, and maybeAddImport becomes a deprecated
one-line delegation to it. bindModule's own options (binding, esmForm)
collapse into AddImportOptions' existing preferredName/member fields.

Reorganizes src/javascript around the new entry point: binding.ts holds
maybeBind, moduleBindings, isAmdBlock, maybeRemoveImport and the deprecated
maybeAddImport shim; amd.ts absorbs everything AMD-specific from the
now-deleted bind-module.ts and remove-amd-dependency.ts; add-import.ts's
maybeAddImport engine is renamed bindImport and stays the ESM/CommonJS lane
maybeBind delegates to. change-import.ts's four maybeAddImport call sites
move to maybeBind, gaining AMD support for free.
