+++ rewrite-javascript/rewrite/test/javascript/add-import.test.ts
@@ -2954,10 +2954,13 @@ describe('AddImport visitor', () => {
 
         test('a name bound through any binding pattern is occupied without answering for the module', async () => {
             const spec = new RecipeSpec();
-            const bound: string[] = [];
+            const bound: (string | undefined)[] = [];
             spec.recipe = fromVisitor(new class extends JavaScriptVisitor<any> {
                 override async visitJsCompilationUnit(cu: JS.CompilationUnit, p: any): Promise<J | undefined> {
                     bound.push(maybeAddImport(this, {module: 'm', member: 'b', onlyIfReferenced: false}));
+                    // 'other' is already required here, so the file reads as CommonJS; a request
+                    // for its module object (no member) shares maybeBind's refusal to gain an ES
+                    // import there (ADR 0013).
                     bound.push(maybeAddImport(this, {module: 'other', preferredName: 'z', onlyIfReferenced: false}));
                     bound.push(maybeAddImport(this, {module: 'p', member: 'default', preferredName: 'e', onlyIfReferenced: false}));
                     return super.visitJsCompilationUnit(cu, p);
@@ -2974,7 +2977,6 @@ describe('AddImport visitor', () => {
                     `,
                     `
                         import {b as b_1} from 'm';
-                        import z from 'other';
                         import e_1 from 'p';
 
                         const {a: {b}} = require('other');
@@ -2985,7 +2987,7 @@ describe('AddImport visitor', () => {
                 )
             );
 
-            expect(bound).toEqual(['b_1', 'z', 'e_1']);
+            expect(bound).toEqual(['b_1', undefined, 'e_1']);
         });
 
         test('a merged specifier sorts by the name it binds, as its neighbours do', async () => {
